requirements.txt
numpy==2.1.2
tqdm==4.66.3
PyYAML==6.0.1
jsonlines==4.0.0
torch==2.5.0
einops==0.8.0
safetensors==0.4.5
omegaconf==2.3.0
hydra-core==1.3.2
pandas==2.2.3
matplotlib==3.9.0
sentencepiece==0.2.0
datasets==2.18.0
transformers==4.46.1
accelerate==1.0.1
wandb==0.17.4
tree-sitter==0.22.3
tree-sitter-python==0.23.2
flash-attn==2.5.9.post1
